Weather dataset schemas, data and queries
In the 'weather_dataset_schema' repository, the data model is available in 'weatherdataset_model.ttl'. Different weather parameters and features of interest are defined in the 'meteofrance_vocabulaire'. An example of a WMO code is defined in 'WMO_thesaurus', other WMO code will be defined soon.

In weather_dataset repository, we have observations of the period of august 2021. All informations about stations is available in 'meteofrance_stations.ttl'.

Examples of queries are defined in sparql_examples repository.